Move commercial installer into its own file
